\name{example.spec}
\alias{example.spec}
\alias{example.meta}
\docType{data}
\title{Data for Example Sea Lion Blubber 2010 Mass Spectral Library}
\description{Sea lion blubber extract analyzed by GCxGC-TOF. Electron impact ionization.}
\usage{example.spec}
\format{\code{example.spec} is a data frame containing the mass
  spectra. \code{example.meta} is a data frame containing the metadata.
}  
\source{Blubber from a 4+ year old male sea lion (\emph{Zalophus californianus}) stranded in southern California around the year 2000 was obtained from California State University Long Beach.

The blubber sample was prepared using pressurized liquid extraction (PLE) with dichloromethane, followed by removal of the majority of the lipids by gel permeation chromatography (GPC), and fractionation by polarity using silica solid phase extraction (SPE). All compounds in the library were detected in the SPE hexane fraction, except for bisphenol A, which was detected in the hexane:dichloromethane 1:1 (volume/volume) fraction.   

The extract was analyzed using comprehensive two dimensional gas chromatography with time of flight mass spectrometry (GCxGC-TOF MS, LECO Pegasus 4D) in electron impact mode. The TOF mass defect correction was not set. The spectra were deconvoluted using the LECO ChromaTOF software.

Spectra were identified by either manual interpretation or spectral matching to the NIST library (MS Search 2.0, 2008). Retention time matches to authentic standards were not made.

The \emph{m/z} labels in the mass spectrum refer to the most abundant peak in the cluster. The peak \emph{m/z} values listed in the table refer to the monoisotopic peak in the cluster. \dQuote{M} refers to the molecular ion.

}
\author{Eunha Hoh and Nathan G. Dodder}
\keyword{datasets} 

